refactor: remove legacy RouterPodProvider
Named netns is the only router architecture — the CRI+PID-based
RouterPodProvider and RouterPod types are no longer used. Remove them,
keeping only the shared helpers (routerPodForNode, PodIsReady) that
RouterNamedNSProvider still depends on.
